+++ pcs/usage.py
@@ -790,6 +790,19 @@ Commands:
     delete <stonith id>
         Remove stonith id from configuration.
 
+    enable <stonith id> [--wait[=n]]
+        Allow the cluster to use the stonith device. If --wait is specified, pcs
+        will wait up to 'n' seconds for the stonith device to start and then
+        return 0 if the stonith device is started, or 1 if the stonith device
+        has not yet started. If 'n' is not specified it defaults to 60 minutes.
+
+    disable <stonith id> [--wait[=n]]
+        Attempt to stop the stonith device if it is running and disallow the
+        cluster to use it. If --wait is specified, pcs will wait up to 'n'
+        seconds for the stonith device to stop and then return 0 if the stonith
+        device is stopped or 1 if the stonith device has not stopped.  If 'n' is
+        not specified it defaults to 60 minutes.
+
     cleanup [<stonith id>] [--node <node>]
         Make the cluster forget the operation history of the stonith device and
         re-detect its current state. This can be useful to purge knowledge of